+++ src/main/java/com/wenxin2/marioverse/datagen/BlockLootTableGen.java
@@ -98,6 +98,9 @@ public class BlockLootTableGen extends LootTableProvider {
                     else if (block == BlockRegistry.FUNGAL_STONE.get())
                         this.add(block, silkTouchBlock -> this.createSingleItemTableWithSilkTouch(silkTouchBlock,
                                 BlockRegistry.FUNGAL_COBBLESTONE.get()));
+                    else if (block == BlockRegistry.SHROOMGRASS_BLOCK.get())
+                        this.add(block, silkTouchBlock -> this.createSingleItemTableWithSilkTouch(silkTouchBlock,
+                                BlockRegistry.SHROOMSOIL.get()));
                     else if (block == BlockRegistry.MUSHROOT_LEAVES.get())
                         this.add(block, this.createLeavesDrops(BlockRegistry.MUSHROOT_LEAVES.get(),
                                 BlockRegistry.MUSHROOT_SAPLING.get(), NORMAL_LEAVES_SAPLING_CHANCES));