+++ src/main/java/com/wenxin2/marioverse/event_handlers/RegistryEventHandlers.java
@@ -554,6 +554,11 @@ public class RegistryEventHandlers {
                 16, 10, 0.2F));
 
         genericTrades.add((entity, random) -> new MerchantOffer(
+                new ItemCost(Items.EMERALD, 3),
+                new ItemStack(BlockRegistry.SHROOMGRASS_BLOCK, 3),
+                16, 10, 0.2F));
+
+        genericTrades.add((entity, random) -> new MerchantOffer(
                 new ItemCost(Items.EMERALD, 2),
                 new ItemStack(BlockRegistry.RED_TRAMPOLINE_CAP, 1),
                 16, 10, 0.2F));